wip: add build command (default)
