+++ source/commands/index.ts
@@ -1,3 +1,4 @@
+export {build} from './build.js';
 export {init} from './init.js';
 export {tidy} from './tidy.js';
 export {watch} from './watch.js';
